Simplified digest in progress check
Angular checks for a digest in progress in this way rather than checking the string '$digest'.
https://github.com/angular/angular.js/blob/65ec0ab70a506b4b52fadecf444a7146d8d425e8/src/ng/http.js#L953

This is a bit more robust because Angular could change the string value ('$digest' to '$$digest' for instance) which would break the check.

Also, the check for rootScope and childScope phase is not necessary. If no digest is occurring on the childScope, it is safe to trigger a childScope digest.